refactor(python): Rename `cfg` module to `config`
